+++ src/main/java/ggfab/ComponentRecipeLoader.java
@@ -36,7 +36,7 @@ class ComponentRecipeLoader implements Runnable {
                 ItemList.Automation_ChestBuffer_LuV.get(1L))
             .fluidInputs(
                 MaterialUtils.anyFluid(Materials.Indalloy140, 9 * INGOTS),
-                MaterialLibAPI.getFluidStack(Materials.Lubricant, FluidShapes.fluidLiquid, (int) (2_000)))
+                MaterialLibAPI.getFluidStack(Materials.Lubricant, FluidShapes.fluidLiquid, 2_000))
             .itemOutputs(GGItemList.AdvAssLine.get(1L))
             .eut(TierEU.RECIPE_LuV)
             .duration(3 * MINUTES)
@@ -47,7 +47,7 @@ class ComponentRecipeLoader implements Runnable {
                 ItemList.Hatch_Input_Bus_IV.get(1L),
                 ItemList.Emitter_IV.get(1L),
                 ItemList.Sensor_IV.get(1L),
-                MaterialLibAPI.getStack(Materials.Enderium, Shapes.plateDense, (int) (1)))
+                MaterialLibAPI.getStack(Materials.Enderium, Shapes.plateDense, 1))
             .circuit(12)
             .itemOutputs(GGItemList.LinkedInputBus.get(1L))
             .fluidInputs(
