+++ src/main/java/goodgenerator/loader/NeutronActivatorLoader.java
@@ -21,34 +21,29 @@ public class NeutronActivatorLoader {
         GTValues.RA.stdBuilder()
             .fluidInputs(
                 MaterialLibAPI
-                    .getFluidStack(Materials.ThoriumBasedLiquidFuelExcitedState, FluidShapes.fluidLiquid, (int) (200)))
+                    .getFluidStack(Materials.ThoriumBasedLiquidFuelExcitedState, FluidShapes.fluidLiquid, 200))
             .fluidOutputs(
-                MaterialLibAPI
-                    .getFluidStack(Materials.ThoriumBasedLiquidFuelDepleted, FluidShapes.fluidLiquid, (int) (200)))
+                MaterialLibAPI.getFluidStack(Materials.ThoriumBasedLiquidFuelDepleted, FluidShapes.fluidLiquid, 200))
             .duration(8 * MINUTES + 20 * SECONDS)
             .eut(0)
             .metadata(NKE_RANGE, computeRangeNKE(700, 500))
             .addTo(neutronActivatorRecipes);
         GTValues.RA.stdBuilder()
             .itemInputs(GTUtility.copyAmount(0, MaterialLibAPI.getStack(Materials.Tungsten, Shapes.plate, 1)))
-            .fluidInputs(
-                MaterialLibAPI.getFluidStack(Materials.UraniumBasedLiquidFuel, FluidShapes.fluidLiquid, (int) (100)))
+            .fluidInputs(MaterialLibAPI.getFluidStack(Materials.UraniumBasedLiquidFuel, FluidShapes.fluidLiquid, 100))
             .fluidOutputs(
                 MaterialLibAPI
-                    .getFluidStack(Materials.UraniumBasedLiquidFuelExcitedState, FluidShapes.fluidLiquid, (int) (100)))
+                    .getFluidStack(Materials.UraniumBasedLiquidFuelExcitedState, FluidShapes.fluidLiquid, 100))
             .duration(4 * SECONDS)
             .eut(0)
             .metadata(NKE_RANGE, computeRangeNKE(550, 450))
             .addTo(neutronActivatorRecipes);
         GTValues.RA.stdBuilder()
             .itemInputs(GTUtility.copyAmount(0, MaterialLibAPI.getStack(Materials.Tritanium, Shapes.plate, 1)))
-            .fluidInputs(
-                MaterialLibAPI.getFluidStack(Materials.PlutoniumBasedLiquidFuel, FluidShapes.fluidLiquid, (int) (100)))
+            .fluidInputs(MaterialLibAPI.getFluidStack(Materials.PlutoniumBasedLiquidFuel, FluidShapes.fluidLiquid, 100))
             .fluidOutputs(
-                MaterialLibAPI.getFluidStack(
-                    Materials.PlutoniumBasedLiquidFuelExcitedState,
-                    FluidShapes.fluidLiquid,
-                    (int) (100)))
+                MaterialLibAPI
+                    .getFluidStack(Materials.PlutoniumBasedLiquidFuelExcitedState, FluidShapes.fluidLiquid, 100))
             .duration(4 * SECONDS)
             .eut(0)
             .metadata(NKE_RANGE, computeRangeNKE(600, 500))
@@ -57,10 +52,9 @@ public class NeutronActivatorLoader {
         GTValues.RA.stdBuilder()
             .itemInputs(ItemList.Tesseract.get(1))
             .fluidInputs(
-                MaterialLibAPI.getFluidStack(Materials.NaquadahBasedLiquidFuelMkV, FluidShapes.fluidLiquid, (int) (64)))
+                MaterialLibAPI.getFluidStack(Materials.NaquadahBasedLiquidFuelMkV, FluidShapes.fluidLiquid, 64))
             .fluidOutputs(
-                MaterialLibAPI
-                    .getFluidStack(Materials.NaquadahBasedLiquidFuelMkVDepleted, FluidShapes.fluidLiquid, (int) (64)))
+                MaterialLibAPI.getFluidStack(Materials.NaquadahBasedLiquidFuelMkVDepleted, FluidShapes.fluidLiquid, 64))
             .itemOutputs(ItemList.EnergisedTesseract.get(1))
             .duration(16400 * SECONDS)
             .eut(0)
@@ -70,11 +64,10 @@ public class NeutronActivatorLoader {
         GTValues.RA.stdBuilder()
             .itemInputs(ItemList.Tesseract.get(1))
             .fluidInputs(
-                MaterialLibAPI
-                    .getFluidStack(Materials.NaquadahBasedLiquidFuelMkVI, FluidShapes.fluidLiquid, (int) (64)))
+                MaterialLibAPI.getFluidStack(Materials.NaquadahBasedLiquidFuelMkVI, FluidShapes.fluidLiquid, 64))
             .fluidOutputs(
                 MaterialLibAPI
-                    .getFluidStack(Materials.NaquadahBasedLiquidFuelMkVIDepleted, FluidShapes.fluidLiquid, (int) (64)))
+                    .getFluidStack(Materials.NaquadahBasedLiquidFuelMkVIDepleted, FluidShapes.fluidLiquid, 64))
             .itemOutputs(ItemList.EnergisedTesseract.get(1))
             .duration(24600 * SECONDS)
             .eut(0)
